# OrderManagement.Api/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Order Management API",
    "version": "1.0",
    "description": "Multi-tenant restaurant order management platform for Malaysia",
    "contact": {
      "name": "Order Management Support",
      "email": "support@ordermanagement.local"
    }
  },
  "servers": [
    {
      "url": "http://localhost:5073",
      "description": "Local Development Server (HTTP)"
    },
    {
      "url": "https://localhost:7118",
      "description": "Local Development Server (HTTPS - requires SSL verification disabled)"
    }
  ],
  "components": {
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Enter JWT Bearer token only"
      }
    },
    "schemas": {
...
      "CreateOrderLineRequest": {
        "type": "object",
        "required": ["menuItemId", "name", "price", "quantity"],
        "properties": {
          "menuItemId": {
            "type": "string",
            "format": "uuid",
            "example": "550e8400-e29b-41d4-a716-446655440000"
          },
          "name": {
            "type": "string",
            "example": "Dosa"
          },
          "price": {
            "type": "number",
            "format": "decimal",
            "example": 12.50
          },
          "quantity": {
            "type": "integer",
            "example": 2
          }
        }
      },
      "CreateOrderRequest": {
        "type": "object",
        "required": ["isTakeAway", "tableNumber", "items"],
        "properties": {
          "isTakeAway": {
            "type": "boolean",
            "example": false
          },
          "tableNumber": {
            "type": "string",
            "example": "T-05"
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CreateOrderLineRequest"
            }
          }
        }
      },
      "Order": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "example": "550e8400-e29b-41d4-a716-446655440010"
          },
          "tenantId": {
            "type": "string",
            "format": "uuid",
            "example": "550e8400-e29b-41d4-a716-446655440002"
          },
          "branchId": {
            "type": "string",
            "format": "uuid",
            "example": "550e8400-e29b-41d4-a716-446655440001"
          },
          "orderNumber": {
            "type": "string",
            "example": "OM-20241215143022"
          },
          "status": {
            "type": "string",
            "enum": ["Draft", "Submitted", "InPreparation", "Ready", "Delivered", "Cancelled", "Paid"],
            "example": "Submitted"
          },
          "subtotal": {
            "type": "number",
            "format": "decimal",
            "example": 25.00
          },
          "tax": {
            "type": "number",
            "format": "decimal",
            "example": 1.50
          },
          "serviceCharge": {
            "type": "number",
            "format": "decimal",
            "example": 2.50
          },
          "total": {
            "type": "number",
            "format": "decimal",
            "example": 29.00
          },
          "isTakeAway": {
            "type": "boolean",
            "example": false
          },
          "tableNumber": {
            "type": "string",
            "example": "T-05"
          },
          "lines": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "menuItemId": {
                  "type": "string",
                  "format": "uuid"
                },
                "name": {
                  "type": "string"
                },
                "price": {
                  "type": "number",
                  "format": "decimal"
                },
                "quantity": {
                  "type": "integer"
                },
                "total": {
                  "type": "number",
                  "format": "decimal"
                }
              }
            }
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "example": "2024-12-15T14:30:22Z"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          }
        }
      },
...
    }
  },
  "paths": {
...
    "/api/v1/Orders": {
      "post": {
        "tags": ["Orders"],
        "summary": "Create a new order",
        "description": "Create a new order with menu items",
        "operationId": "CreateOrder",
        "security": [{"Bearer": []}],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateOrderRequest"
              },
              "example": {
                "isTakeAway": false,
                "tableNumber": "T-05",
                "items": [
                  {
                    "menuItemId": "550e8400-e29b-41d4-a716-446655440000",
                    "name": "Dosa",
                    "price": 12.50,
                    "quantity": 2
                  },
                  {
                    "menuItemId": "550e8400-e29b-41d4-a716-446655440003",
                    "name": "Nasi Lemak",
                    "price": 8.50,
                    "quantity": 1
                  }
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Order created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Order"
                },
                "example": {
                  "id": "550e8400-e29b-41d4-a716-446655440010",
                  "tenantId": "550e8400-e29b-41d4-a716-446655440002",
                  "branchId": "550e8400-e29b-41d4-a716-446655440001",
                  "orderNumber": "OM-20241215143022",
                  "status": "Submitted",
                  "subtotal": 33.50,
                  "tax": 2.01,
                  "serviceCharge": 3.35,
                  "total": 38.86,
                  "isTakeAway": false,
                  "tableNumber": "T-05",
                  "lines": [
                    {
                      "menuItemId": "550e8400-e29b-41d4-a716-446655440000",
                      "name": "Dosa",
                      "price": 12.50,
                      "quantity": 2,
                      "total": 25.00
                    },
                    {
                      "menuItemId": "550e8400-e29b-41d4-a716-446655440003",
                      "name": "Nasi Lemak",
                      "price": 8.50,
                      "quantity": 1,
                      "total": 8.50
                    }
                  ],
                  "createdAt": "2024-12-15T14:30:22Z"
                }
              }
            }
          },
          "400": {
            "description": "Bad request - Branch context required"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/api/v1/Orders/{id}": {
      "get": {
        "tags": ["Orders"],
        "summary": "Get order by ID",
        "description": "Retrieve a specific order by its ID",
        "operationId": "GetOrder",
        "security": [{"Bearer": []}],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "example": "550e8400-e29b-41d4-a716-446655440010"
          }
        ],
        "responses": {
          "200": {
            "description": "Order found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Order"
                }
              }
            }
          },
          "404": {
            "description": "Order not found"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/api/v1/Orders/{id}/status": {
      "patch": {
        "tags": ["Orders"],
        "summary": "Update order status",
        "description": "Update the status of an order",
        "operationId": "UpdateOrderStatus",
        "security": [{"Bearer": []}],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "example": "550e8400-e29b-41d4-a716-446655440010"
          },
          {
            "name": "status",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "enum": ["Draft", "Submitted", "InPreparation", "Ready", "Delivered", "Cancelled", "Paid"]
            },
            "example": "InPreparation"
          }
        ],
        "responses": {
          "204": {
            "description": "Order status updated successfully"
          },
          "404": {
            "description": "Order not found"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/api/v1/Orders/qr": {
      "post": {
        "tags": ["Orders"],
        "summary": "Generate QR ordering session",
        "description": "Generate a QR code session for customer self-ordering",
        "operationId": "GenerateQrSession",
        "security": [{"Bearer": []}],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "string"
              },
              "example": "T-05"
            }
          }
        },
        "responses": {
          "200": {
            "description": "QR session generated",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string",
                      "example": "QR-20241215143022-T-05"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request - Branch context required"
          }
        }
      }
    },
...
    "/api/v1/Reports/sales": {
      "get": {
        "tags": ["Reports"],
        "summary": "Get sales report",
        "description": "Get sales summary for a date range",
        "operationId": "GetSalesReport",
        "security": [{"Bearer": []}],
        "parameters": [
          {
            "name": "from",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date-time"
            },
            "example": "2024-12-01T00:00:00Z"
          },
          {
            "name": "to",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date-time"
            },
            "example": "2024-12-15T23:59:59Z"
          }
        ],
        "responses": {
          "200": {
            "description": "Sales report",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                },
                "example": {
                  "totalSales": 12500.50,
                  "totalOrders": 150,
                  "averageOrderValue": 83.34,
                  "period": {
                    "from": "2024-12-01T00:00:00Z",
                    "to": "2024-12-15T23:59:59Z"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/api/v1/Reports/inventory": {
      "get": {
        "tags": ["Reports"],
        "summary": "Get inventory report",
        "description": "Get inventory forecast report",
        "operationId": "GetInventoryReport",
        "security": [{"Bearer": []}],
        "responses": {
          "200": {
            "description": "Inventory report",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                },
                "example": {
                  "lowStockItems": 5,
                  "outOfStockItems": 2,
                  "totalItems": 50
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/api/v1/Reports/subscription": {
      "get": {
        "tags": ["Reports"],
        "summary": "Get subscription usage report",
        "description": "Get subscription usage statistics (SuperAdmin/Admin only)",
        "operationId": "GetSubscriptionUsage",
        "security": [{"Bearer": []}],
        "responses": {
          "200": {
            "description": "Subscription usage report",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                },
                "example": {
                  "currentPlan": "Basic",
                  "branchesUsed": 3,
                  "branchesLimit": 5,
                  "usersUsed": 25,
                  "usersLimit": 50,
                  "ordersThisMonth": 1250
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden - SuperAdmin/Admin only"
          }
        }
      }
    }
  },
...
}